+++ garage/spaces/__init__.py
@@ -2,6 +2,6 @@ from garage.spaces.base import Space
 from garage.spaces.box import Box
 from garage.spaces.dict import Dict
 from garage.spaces.discrete import Discrete
-from garage.spaces.product import Product
+from garage.spaces.tuple import Tuple
 
-__all__ = ["Space", "Box", "Dict", "Discrete", "Product"]
+__all__ = ["Space", "Box", "Dict", "Discrete", "Tuple"]
