+++ garage/tf/spaces/__init__.py
@@ -3,6 +3,6 @@
 from garage.tf.spaces.box import Box
 from garage.tf.spaces.dict import Dict
 from garage.tf.spaces.discrete import Discrete
-from garage.tf.spaces.product import Product
+from garage.tf.spaces.tuple import Tuple
 
-__all__ = ["Box", "Dict", "Discrete", "Product"]
+__all__ = ["Box", "Dict", "Discrete", "Tuple"]
