+++ garage/theano/spaces/__init__.py
@@ -3,6 +3,6 @@
 from garage.theano.spaces.box import Box
 from garage.theano.spaces.dict import Dict
 from garage.theano.spaces.discrete import Discrete
-from garage.theano.spaces.product import Product
+from garage.theano.spaces.tuple import Tuple
 
-__all__ = ["Box", "Dict", "Discrete", "Product"]
+__all__ = ["Box", "Dict", "Discrete", "Tuple"]
