+++ tests/garage/tf/spaces/test_tuple.py
@@ -1,13 +1,13 @@
 import pickle
 import unittest
 
-from garage.spaces.product import Product
-from garage.theano.spaces import Discrete
+from garage.tf.spaces import Discrete
+from garage.tf.spaces.tuple import Tuple
 
 
-class TestProduct(unittest.TestCase):
+class TestTuple(unittest.TestCase):
     def test_pickleable(self):
-        obj = Product((Discrete(3), Discrete(2)))
+        obj = Tuple((Discrete(3), Discrete(2)))
         round_trip = pickle.loads(pickle.dumps(obj))
         assert round_trip
         assert round_trip.components == obj.components
