+++ tests/garage/theano/spaces/test_spaces.py
@@ -4,18 +4,18 @@ import numpy as np
 
 from garage.theano.spaces import Box
 from garage.theano.spaces import Discrete
-from garage.theano.spaces import Product
+from garage.theano.spaces import Tuple
 
 
 class TestSpaces(unittest.TestCase):
-    def test_product_space(self):
-        _ = Product([Discrete(3), Discrete(2)])
-        product_space = Product(Discrete(3), Discrete(2))
-        sample = product_space.sample()
-        assert product_space.contains(sample)
+    def test_tuple_space(self):
+        _ = Tuple([Discrete(3), Discrete(2)])
+        tuple_space = Tuple(Discrete(3), Discrete(2))
+        sample = tuple_space.sample()
+        assert tuple_space.contains(sample)
 
-    def test_product_space_unflatten_n(self):
-        space = Product([Discrete(3), Discrete(3)])
+    def test_tuple_space_unflatten_n(self):
+        space = Tuple([Discrete(3), Discrete(3)])
         np.testing.assert_array_equal(
             space.flatten((2, 2)),
             space.flatten_n([(2, 2)])[0])
